+++ quetzal/app/api/data/tasks.py
@@ -8,13 +8,15 @@ from urllib.parse import urlparse
 from flask import current_app
 from sqlalchemy import func, types
 from sqlalchemy.sql.ddl import CreateSchema
+from sqlalchemy.sql import literal
+from sqlalchemy.sql.functions import coalesce
 from sqlalchemy.dialects.postgresql import UUID
 
 from quetzal.app import celery, db
 from quetzal.app.api.exceptions import Conflict, EmptyCommit, WorkerException
 from quetzal.app.helpers.google_api import get_client, get_bucket, get_data_bucket
 from quetzal.app.helpers.sql import CreateTableAs, DropSchemaIfExists, GrantUsageOnSchema
-from quetzal.app.models import Family, FileState, Metadata, Workspace, WorkspaceState
+from quetzal.app.models import Family, FileState, Metadata, QueryDialect, Workspace, WorkspaceState
 
 
 logger = logging.getLogger(__name__)
@@ -286,7 +288,6 @@ def _delete_local_data_bucket(url):
 
 @celery.task()
 def scan_workspace(wid):
-    logger.info('Scanning workspace %s...', wid)
 
     # Get the workspace object and verify preconditions
     workspace = Workspace.query.get(wid)
@@ -296,8 +297,89 @@ def scan_workspace(wid):
     if workspace.state != WorkspaceState.SCANNING:
         raise WorkerException('Workspace was not on the expected state')
 
+    schema_name = workspace.make_schema_name()
+    _scan_create_table_views(workspace, schema_name)
+    _scan_create_json_views(workspace, schema_name)
+
+    # Update the workspace object to have the correct schema and state
+    workspace.pg_schema_name = schema_name
+    workspace.state = WorkspaceState.READY
+    db.session.add(workspace)
+
+    # Commit all changes
+    db.session.commit()
+
+
+def _new_schema(workspace, name, suffix):
+    # Drop previous schema
+    if workspace.pg_schema_name is not None:
+        old_schema = workspace.pg_schema_name + suffix
+        db.session.execute(DropSchemaIfExists(old_schema, cascade=True))
+    # create postgres schema
+    new_name = name + suffix
+    db.session.execute(CreateSchema(new_name))
+    return new_name
+
+
+def _scan_create_json_views(workspace, schema_name):
+    logger.info('Scanning workspace %s to create json views...', workspace.id)
+
     # The scanning task consists on the following procedure:
     # 1. create a new database directory (aka schema in postgres)
+    #    also, drop the previous one if it existed
+    # 2. get the workspace metadata
+    # 3. for each family f
+    # 3.1. create a subquery that selects only the latest metadata of family f
+    # 4. join all family queries with the appropriate aliases so that the
+    #    column names correspond to the family name
+    # 4. grant permissions on the directory
+    # 5. drop previous directory if any
+
+    # Drop previous schema and create a new one
+    suffix = f'_{QueryDialect.POSTGRESQL_JSON.value}'
+    new_schema = _new_schema(workspace, schema_name, suffix)
+
+    # Extract metadata per family
+    workspace_metadata = workspace.get_metadata()
+    subqueries = []
+    sorted_families = sorted(workspace.families,
+                             key=lambda fam: (-1, fam.name) if fam.name == 'base' else (+1, fam.name))
+    for family in sorted_families:
+        sub = (
+            workspace_metadata
+            .filter(Family.name == family.name)
+            .subquery(name=family.name)
+        )
+        subqueries.append(sub)
+
+    # Make a joined table of all metadata and set the json column to the name of the family
+    q1 = subqueries.pop(0)  # the first one is always the base family, due to the sort done before
+    joined_query = db.session.query(q1)
+    columns = [q1.c.metadata_id_file.label('id'), q1.c.metadata_json.label('base')]
+    for q2 in subqueries:
+        joined_query = joined_query.outerjoin((q2, q1.c.metadata_id_file == q2.c.metadata_id_file))
+        # Here, we are coalescing to set an empty dict to files that do not
+        # have an entry for this particular family. This does not apply to the
+        # base family because the base family is always present
+        columns.append(coalesce(q2.c.metadata_json, literal({}, types.JSON)).label(q2.name))
+        q1 = q2
+
+    master_query = joined_query.with_entities(*columns).subquery()
+
+    family_table_name = f'{new_schema}.metadata'
+    create_table_statement = CreateTableAs(family_table_name, master_query)
+    db.session.execute(create_table_statement)
+
+    # Set permissions on readonly user to the schema contents
+    db.session.execute(GrantUsageOnSchema(new_schema, 'db_ro_user'))
+
+
+def _scan_create_table_views(workspace, schema_name):
+    logger.info('Scanning workspace %s to create table views...', workspace.id)
+
+    # The scanning task consists on the following procedure:
+    # 1. create a new database directory (aka schema in postgres)
+    #    also, drop the previous one if it existed
     # 2. get the workspace metadata
     # 3. for each family f
     # 3.1. determine the type of each column (this is currently not very smart)
@@ -307,10 +389,9 @@ def scan_workspace(wid):
     # 4. grant permissions on the directory
     # 5. drop previous directory if any
 
-    # create postgres schema
-    old_schema = workspace.pg_schema_name
-    new_schema = workspace.make_schema_name()
-    db.session.execute(CreateSchema(new_schema))
+    # Drop previous schema and create a new one
+    suffix = f'_{QueryDialect.POSTGRESQL.value}'
+    new_schema = _new_schema(workspace, schema_name, suffix)
 
     # 2. For each family
     workspace_metadata = workspace.get_metadata()
@@ -350,18 +431,6 @@ def scan_workspace(wid):
 
     # Set permissions on readonly user to the schema contents
     db.session.execute(GrantUsageOnSchema(new_schema, 'db_ro_user'))
-
-    # Drop previous schema
-    if old_schema is not None:
-        db.session.execute(DropSchemaIfExists(old_schema, cascade=True))
-
-    # Update the workspace object to have the correct schema and state
-    workspace.pg_schema_name = new_schema
-    workspace.state = WorkspaceState.READY
-    db.session.add(workspace)
-
-    # Commit all changes
-    db.session.commit()
 
 
 @celery.task()
@@ -616,10 +685,16 @@ def merge(ancestor, theirs, mine):
 
 def _update_global_views():
     # TODO: protect with a database lock
+    _update_global_table_views(f'global_views_{QueryDialect.POSTGRESQL.value}')
+    _update_global_json_views(f'global_views_{QueryDialect.POSTGRESQL_JSON.value}')
+
+
+def _update_global_table_views(schema_name):
+    logger.info('Updating global table views')
 
     # Create postgres schema
-    db.session.execute(DropSchemaIfExists('global_views', cascade=True))
-    db.session.execute(CreateSchema('global_views'))
+    db.session.execute(DropSchemaIfExists(schema_name, cascade=True))
+    db.session.execute(CreateSchema(schema_name))
 
     # Get all the known families
     families = Family.query.filter(Family.fk_workspace_id.is_(None)).distinct(Family.name)
@@ -654,11 +729,59 @@ def _update_global_views():
             .with_entities(*columns)
             .subquery()
         )
-        family_table_name = f'global_views.{family.name}'
+        family_table_name = f'{schema_name}.{family.name}'
         create_table_statement = CreateTableAs(family_table_name, create_table_query)
         db.session.execute(create_table_statement)
 
         # TODO: create an index on the id column
 
     # Set permissions on readonly user to the schema contents
-    db.session.execute(GrantUsageOnSchema('global_views', 'db_ro_user'))
+    db.session.execute(GrantUsageOnSchema(schema_name, 'db_ro_user'))
+
+
+def _update_global_json_views(schema_name):
+    logger.info('Updating global json views')
+
+    # Create postgres schema
+    db.session.execute(DropSchemaIfExists(schema_name, cascade=True))
+    db.session.execute(CreateSchema(schema_name))
+
+    # Get all the known families
+    families = Family.query.filter(Family.fk_workspace_id.is_(None)).distinct(Family.name)
+    # This is the metadata entries related to the latest global families
+    global_metadata = Metadata.get_latest_global()
+
+    # Extract metadata per family
+    subqueries = []
+    sorted_families = sorted(families,
+                             key=lambda fam: (-1, fam.name) if fam.name == 'base' else (+1, fam.name))
+    for family in sorted_families:
+        sub = (
+            global_metadata
+            .filter(Family.name == family.name)
+            .subquery(name=family.name)
+        )
+        subqueries.append(sub)
+
+    # Make a joined table of all metadata and set the json column to the name of the family
+    q1 = subqueries.pop(0)  # the first one is always the base family, due to the sort done before
+    joined_query = db.session.query(q1)
+    columns = [q1.c.id_file.label('id'), q1.c.json.label('base')]
+    for q2 in subqueries:
+        joined_query = joined_query.outerjoin((q2, q1.c.id_file == q2.c.id_file))
+        # Here, we are coalescing to set an empty dict to files that do not
+        # have an entry for this particular family. This does not apply to the
+        # base family because the base family is always present
+        columns.append(coalesce(q2.c.json, literal({}, types.JSON)).label(q2.name))
+        q1 = q2
+
+    master_query = joined_query.with_entities(*columns).subquery()
+
+    family_table_name = f'{schema_name}.metadata'
+    create_table_statement = CreateTableAs(family_table_name, master_query)
+    db.session.execute(create_table_statement)
+
+    # TODO: create an index on the id column
+
+    # Set permissions on readonly user to the schema contents
+    db.session.execute(GrantUsageOnSchema(schema_name, 'db_ro_user'))